convert code URN to code URL
